Final version: Code used in BunnyBots 2015 competition.
This commit contains the latest version of the code used in the
BunnyBots 2015 competition. Following are changes made before and during
the competition:
* PIDMecanumDrive and all of its commands were deprecated.
* Autonomous was implemented: First, the claw lowers down, grabs the
bunny, and rises back up. Then, the robot drives forward, secures the
football, and drives to the middle of the field.
* A Delay command was implemented for timing during autonomous.
* The DriveFromPresets command allows for specification of robot
movement through simple parameters.
* The LowerClaw command, well, lowers the claw until the limit switch is
hit.
* The RaiseClaw command does the opposite of the LowerClaw command.
* The SetClawGrabSpeed command is used for setting the speed of the
claw.
* Some port values in RobotMap.cpp were incorrect. Those were changed.
* Other miscellaneous changes.